% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateByDate.R
\name{aggregateByDate}
\alias{aggregateByDate}
\title{aggregateByDate}
\usage{
aggregateByDate(mat, dates, format = "\%m")
}
\arguments{
\item{format}{"%Y-%m-%d": year, month, daty}
}
\description{
aggregateByDate
}
