% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsreg1_adj.R
\name{attribute_change}
\alias{attribute_change}
\title{attribute_change}
\usage{
attribute_change(plsr, l_trend, nyear = 1)
}
\arguments{
\item{plsr}{plsr object}

\item{l_trend}{list(trend, pvalue), trend is a data.frame}
}
\description{
attribute_change
}
