% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_matrix.R
\name{corr_matrix}
\alias{corr_matrix}
\title{corr_matrix}
\usage{
corr_matrix(X, Y)
}
\arguments{
\item{X}{[time, grid], the dimension of X and Y should be same.}

\item{Y}{[time, grid]}
}
\description{
Correlation of each pixel.
}
