% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mktrend_rcpp.R
\name{mkTrend_rcpp}
\alias{mkTrend_rcpp}
\title{Modified Mann Kendall}
\usage{
mkTrend_rcpp(x, ci = 0.95, IsPlot = FALSE)
}
\arguments{
\item{x}{numeric vector}

\item{ci}{critical value of autocorrelation}
}
\value{
\itemize{
\item \code{Z0}   : The original (non corrected) Mann-Kendall test Z statistic. \cr
\item \code{pval0}: The original (non corrected) Mann-Kendall test p-value \cr
\item \code{Z}    : The new Z statistic after applying the correction \cr
\item \code{pval} : Corrected p-value after accounting for serial autocorrelation \cr
\code{N/n*s} Value of the correction factor, representing the quotient of the number
of samples N divided by the effective sample size \code{n*s} \cr
}

slp  : Sen slope, The slope of the (linear) trend according to Sen test
}
\description{
If valid observations <= 5, NA will be returned.
}
\details{
Dongdong Kong
}
\examples{
x <- c(4.81,4.17,4.41,3.59,5.87,3.83,6.03,4.89,4.32,4.69)
r_cpp <- mkTrend_rcpp(x, IsPlot = TRUE)
r <- mkTrend(x)

}
\references{
fume package
}
