% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.barchart2.R
\name{panel.addbarchart_old}
\alias{panel.addbarchart_old}
\title{add barchart of every cluster}
\usage{
panel.addbarchart_old(z, subscripts, cols.avg, showCluster = T, ...)
}
\description{
add barchart of every cluster
}
