% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.barchart2.R
\name{panel.barchart2}
\alias{panel.barchart2}
\title{panel.barchart2}
\usage{
panel.barchart2(z, subscripts, origin.x = 76, origin.y = 26.5,
  tck = 0.2, A = 15, by = 0.6, axis.x.text = TRUE, box.width = by
  * 0.85, col, fontfamily = "Times", border = "transparent",
  ntick = 2, ylab.offset = 2, text.cex = 1, ...)
}
\arguments{
\item{{[type]}}{perc      [description]}

\item{{Number}}{origin.x  [description]}

\item{{Number}}{origin.y  [description]}

\item{{Number}}{tck       [description]}

\item{{Number}}{dx        [description]}

\item{{Number}}{A         [description]}

\item{{Number}}{by        [description]}

\item{{[type]}}{col       Colors}

\item{{[type]}}{box.width [description]}

\item{{Number}}{ntick     [description]}

\item{{Number}}{text.cex  [description]}

\item{{...[type]}}{[description]}
}
\value{
{[type]}              [description]
}
\description{
new panel function for lattice, plot barchart in the bottomleft corner of subplot
}
