% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_phenofit.R
\name{pheno_valid}
\alias{pheno_valid}
\title{pheno_valid}
\usage{
pheno_valid(pheno_sat.grow, pheno_obs.SOS, pheno_obs.EOS)
}
\arguments{
\item{pheno_sat_grow}{A List, with the length of years, the elements of every
list are phenology metrics (This time was 17 matrix, TRS2, TRS5, TR6, DES, GU, ZHANG)}

\item{pheno_sat}{A matrix, dims = \link{year, methods}, same as pheno_sat.grow,}

\item{pheno_obs}{A dataframe,  at least have year and DOY column}
}
\description{
pheno_valid
}
