% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenofit_input.avhrr.R
\name{phenofit_input.avhrr}
\alias{phenofit_input.avhrr}
\title{phenofit_input.avhrr}
\usage{
phenofit_input.avhrr(mat_y, mat_qc, d_date, wmax = 0.8, I_st, st, outfile)
}
\description{
prepare input for phenofit obj
}
