% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenologyTP-package.R
\docType{package}
\name{phenologyTP-package}
\alias{phenologyTP}
\alias{phenologyTP-package}
\title{phenologyTP: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kongdd/phenofit}
  \item Report bugs at \url{https://github.com/kongdd/phenofit/issues}
}

}
\keyword{internal}
