% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsr_attributable.R
\name{plsr_attributable}
\alias{plsr_attributable}
\title{This function only suits for EOS model, Tmin, Tmax, Prec, Srad and SOS as INPUT}
\usage{
plsr_attributable(d, include.fitted = TRUE)
}
\arguments{
\item{d}{Pre-season data.frame returned by \code{\link{get_preseason}}.}
}
\description{
This function only suits for EOS model, Tmin, Tmax, Prec, Srad and SOS as INPUT
}
