% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsreg1.R
\name{plsreg1}
\alias{plsreg1}
\title{plsreg1}
\usage{
plsreg1(predictors, response, comps = 2, crosval = TRUE,
  limq2 = 0.05)
}
\arguments{
\item{crosval}{Logical indicating whether 10 fold cross-validation should be 
performed (TRUE by default). No cross-validation is done, if there is 
missing data or if there are less than 10 observations.}
}
\description{
plsreg1
}
