% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsreg1_adj.R
\name{plsreg1_adj}
\alias{plsreg1_adj}
\title{plsreg1_adj}
\usage{
plsreg1_adj(X, Y, comps = 2, autoVars = TRUE, nminVar = 2,
  minVIP = 0.8)
}
\description{
plsreg1_adj
}
\examples{
\dontrun{
plsreg1_adj(X, Y, comps = 2, autoVars = TRUE, nminVar = 2, minVIP = 0.8)
}
}
