% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenofit_input.avhrr.R
\name{predict_date}
\alias{predict_date}
\title{predict_date}
\usage{
predict_date(d_date, xout)
}
\description{
predict_date
}
\examples{
date_AVHRR <- get_date_AVHRR()
predict_date(date_AVHRR, date_AVHRR$I)
}
