% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pheno.R
\name{read_MCD12Q2}
\alias{read_MCD12Q2}
\title{read MCD12Q2 V5 vegetation phenology}
\usage{
read_MCD12Q2(file)
}
\description{
read MCD12Q2 V5 vegetation phenology
}
\details{
\itemize{
\item \code{Onset_Greenness_Increase1}:	Days since Jan 1, 2000 that correspond to vegetation greenup, mode 1	0	32766
\item \code{Onset_Greenness_Increase2}:	Days since Jan 1, 2000 that correspond to vegetation greenup, mode 2	0	32766
\item \code{Onset_Greenness_Maximum1}:	Days since Jan 1, 2000 that correspond to vegetation maturity, mode 1	0	32766
\item \code{Onset_Greenness_Maximum2}:	Days since Jan 1, 2000 that correspond to vegetation maturity, mode 2	0	32766
\item \code{Onset_Greenness_Decrease1}:	Days since Jan 1, 2000 that correspond to vegetation senescence, mode 1	0	32766
\item \code{Onset_Greenness_Decrease2}:	Days since Jan 1, 2000 that correspond to vegetation senescence, mode 2	0	32766
\item \code{Onset_Greenness_Minimum1}:	Days since Jan 1, 2000 that correspond to vegetation dormancy, mode 1	0	32766
\item \code{Onset_Greenness_Minimum2}:	Days since Jan 1, 2000 that correspond to vegetation dormancy, mode 2
}
}
\references{
https://developers.google.com/earth-engine/datasets/catalog/MODIS_MCD12Q2
}
