% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_pheno.R
\name{read_MCD12Q2_V6}
\alias{read_MCD12Q2_V6}
\title{read MCD12Q2 V5 vegetation phenology}
\usage{
read_MCD12Q2_V6(file)
}
\description{
\enumerate{
\item NumCycles	Total number of valid vegetation cycles with peak in product year	7	0
\item Greenup_1	Date when EVI2 first crossed 15\% of the segment EVI2 amplitude, cycle 1. Days since Jan 1, 1970.	11138	32766
\item Greenup_2	Date when EVI2 first crossed 15\% of the segment EVI2 amplitude, cycle 2. Days since Jan 1, 1970.	11138	32766
\item MidGreenup_1	Date when EVI2 first crossed 50\% of the segment EVI2 amplitude, cycle 1. Days since Jan 1, 1970.	11138	32766
\item MidGreenup_2	Date when EVI2 first crossed 50\% of the segment EVI2 amplitude, cycle 2. Days since Jan 1, 1970.	11138	32766
\item Peak_1	Date when EVI2 reached the segment maximum, cycle 1. Days since Jan 1, 1970.	11138	32766
\item Peak_2	Date when EVI2 reached the segment maximum, cycle 2. Days since Jan 1, 1970.	11138	32766
\item Maturity_1	Date when EVI2 first crossed 90\% of the segment EVI2 amplitude, cycle 1. Days since Jan 1, 1970.	11138	32766
\item Maturity_2	Date when EVI2 first crossed 90\% of the segment EVI2 amplitude, cycle 2. Days since Jan 1, 1970.	11138	32766
\item MidGreendown_1	Date when EVI2 last crossed 50\% of the segment EVI2 amplitude, cycle 1. Days since Jan 1, 1970.	11138	32766
\item MidGreendown_2	Date when EVI2 last crossed 50\% of the segment EVI2 amplitude, cycle 2. Days since Jan 1, 1970.	11138	32766
\item Senescence_1	Date when EVI2 last crossed 90\% of the segment EVI2 amplitude, cycle 1. Days since Jan 1, 1970.	11138	32766
\item Senescence_2	Date when EVI2 last crossed 90\% of the segment EVI2 amplitude, cycle 2. Days since Jan 1, 1970.	11138	32766
\item Dormancy_1	Date when EVI2 last crossed 15\% of the segment EVI2 amplitude, cycle 1. Days since Jan 1, 1970.	11138	32766
\item Dormancy_2	Date when EVI2 last crossed 15\% of the segment EVI2 amplitude, cycle 2. Days since Jan 1, 1970.	11138	32766
}
}
\details{
\itemize{
\item \code{0.15}: Greenup, Dormancy
\item \code{0.5}: MidGreenup, MidGreendown
\item \code{0.9}: Maturity, Senescence
}
}
