% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_VSI.R
\name{vsi_month}
\alias{vsi_month}
\title{vsi_month}
\usage{
vsi_month(full, t1 = FALSE)
}
\arguments{
\item{full}{monthly data.frame \code{[NDVI, t1, Prec, Srad, Tmin, Tavg, Tmax]} input data}
}
\description{
vsi_month
}
