% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsreg1_adj.R
\name{attribute_change}
\alias{attribute_change}
\title{attribute_change}
\usage{
attribute_change(obj, l_trend, nyear = 1, model = "SOS")
}
\arguments{
\item{obj}{plsr object, list(\code{SOS}, \code{NonSOS}, I_grid)}

\item{l_trend}{list(trend, pvalue), trend is a data.frame}
}
\value{
data.table
}
\description{
attribute_change
}
