% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/critical_pcor.R
\name{critical}
\alias{critical}
\alias{critical_pcor}
\alias{critical_pcor2}
\alias{critical_autor}
\title{critical}
\usage{
critical_pcor(n, gn, alpha = c(0.1, 0.05, 0.01))

critical_pcor2(n, gn, alpha = c(0.1, 0.05, 0.01))

critical_autor(n, alpha = c(0.1, 0.05, 0.01))
}
\arguments{
\item{n}{the number of samples}

\item{gn}{the number of given variables}
}
\description{
critical
}
\examples{
critical_pcor(34, 2)
critical_pcor2(34, 2)
}
\references{
[1] Chen Yanguang (2012), Geographical Data analysis with MATLAB. \cr
[2] Kim, S. (2015) ppcor: An R Package for a Fast Calculation to Semi-partial 
Correlation Coefficients. Communications for Statistical Applications and Methods, 
22(6), 665-674. \cr
[3] pcor p.value calculation in ppcor. \cr
}
