% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_preseason.R
\name{get_preseason}
\alias{get_preseason}
\title{get_preseason}
\usage{
get_preseason(Tmin, Tmax, Prec, Srad, SOS, EOS, dates_mete,
  DateRange_pheno = NULL, maxDays = 150)
}
\arguments{
\item{Tmin, }{Tmax Prec, Srad: daily time-series}

\item{SOS, }{EOS: yearly phenological metrics}

\item{dates_mete}{dates of mete}

\item{DateRange_pheno}{range of phenology data}

\item{maxDays}{max length of preseason。}
}
\description{
Get preseason of autumn phenology (EOS).
}
\examples{
\dontrun{
dates <- seq(ymd('19820101'), ymd('20151231'), by = "day")
info_date <- data.table(date = dates, year = year(dates), yday = yday(dates))
get_preseason(Tmin, Tmax, Prec, Srad, SOS, EOS, info_date)
}
}
