% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewise.R
\name{piecewise}
\alias{piecewise}
\title{PieceWise Regression}
\usage{
piecewise(y, year.origin = 1, predict = FALSE)
}
\arguments{
\item{y}{Numeric vector}

\item{year.origin}{first year of y}
}
\description{
PieceWise Regression
}
